refactor: extract IssueExtractionRequestBuilder from IssueExtractionService (#432)
IssueExtractionService combined prompt construction, request payload
assembly, HTTP transport, response parsing, schema repair, and issue
normalization in one 1,100-line service. This is the first slice of the
planned decomposition (ai-task: extract the no-network pieces, builder
first): move the request/payload-building half into a dedicated, network-
free IssueExtractionRequestBuilder.

- New IssueExtractionRequestBuilder owns makeRequest + the prompt assembly
  (makePrompt / makeUserMessageParts / makeScreenshotContentPart / mimeType)
  and the request-side Encodable types + IssueExtractionRequestBudget.
- IssueExtractionService.extractIssues now delegates request construction
  to the builder; transport, parsing, and normalization are unchanged. The
  public service API is unchanged.
- Behavior is byte-compatible (verbatim move). A new characterization test
  suite pins the request shape: the HTTP envelope, model/temperature/
  response_format, the system prompt verbatim, the user-prompt assembly,
  and the screenshot content-part path (image_url keying, exact data URL,
  part ordering, and the budget note) — the last added after build review
  flagged the no-screenshot fixture left the image path unpinned.

Next slices (separate PRs): extract the response parser + schema-repair
mapper, then the issue normalization helpers.

Refs #432
